requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.1.7
django-betterforms==1.2
django-crispy-forms==1.13.0
django-gunicorn==0.1.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
Pillow==8.3.1
platformdirs==2.2.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
